refactor(test): rename spec files
