+++ packages/headless-driver-runner/package.json
@@ -52,7 +52,7 @@
       }
     },
     "testMatch": [
-      "<rootDir>/src/__tests__/**/*Spec.ts"
+      "<rootDir>/src/__tests__/**/*.spec.ts"
     ]
   }
 }
