+++ packages/headless-driver/package.json
@@ -75,7 +75,7 @@
       }
     },
     "testMatch": [
-      "<rootDir>/src/__tests__/**/*Spec.ts"
+      "<rootDir>/src/**/*.spec.ts"
     ],
     "globalSetup": "<rootDir>/src/__tests__/setup.ts",
     "globalTeardown": "<rootDir>/src/__tests__/teardown.ts"
