+++ packages/headless-driver/tsconfig.json
@@ -8,6 +8,7 @@
     "src/**/*.ts"
   ],
   "exclude": [
-    "src/**/__tests__/**/*.ts"
+    "src/**/__tests__/**/*.ts",
+    "src/**/*.spec.ts"
   ]
 }
